Loop Redirects for Canonical Redirection if HTTP Request crosses more than one Reverse Proxies
The Host of the current url of your request, created to perform a match with the canonical url, is not generated respecting checks performed by the method Tools::getHttpHost. This cause always a mismatch if your host is behind two or more proxies.

You can test this problem i.e. in Azure, if your domain points to a FrontDoor Service with Azure and it redirects the request to an internal host having another reverse proxy server(like Nginx, Apache Traffic Server) beyond your Server Web.