Update Cart.php
On a system with dutch NLS the calcultations on line 3250 and 3260 simply did not work. The sum of the calulations was always 0. 
This caused the cart to always show 'free shipping'.
Also, since only one amount is returned there is not much point in calculating both with as without tax.