Bug Fix: Multiple options with a couple in stock
The proposed change will fix a bug where a product has multiple options, with one or more having stock, you can't select an option that does not have stock when you have set the preference to allow out of stock product options to be purchased.

The bug is you have a product with 5 options (S / M / L / XL / XXL) XL is in stock, the others are not in stock.  You have chosen to allow out of stock product to be purchased, because you have one product in stock, when selecting one of the out of stock options the page will reload and go back to the in-stock option as default.  

The bug is the filter that is used for quantity > 0 when the config option of PS_DISP_UNAVAILABLE_ATTR is set.