Fixed free shipping voucher not applying
If you create a voucher with free shipping, without code and also restrict it using carriers, then it will not automatically apply when the customer is in cart page or in the delivery step at checkout. In order for the voucher to apply the customer must manually select carriers or proceed to the payment step. There is also no indication of "Free" next to the carrier name. After the customer proceeds to the payment step everything works fine, but many people might abandon their order if they are expecting to get free shipping and don't see it.