order-final-summary shipping carrier name remove
When the selected carrier delay contains text, that is most likely the widest column. On the other hand the carrier image is most often the company logo.

When the company logo is displayed, is no need to display the name in the 2nd column, as it is already the alternative text of the image. This change re-uses the first column in the summary to display the selected carrier name when no carrier images is present. By that change the carrier-name column can be removed, and the selected_delivery_option.delay column with can be doubled.