CO: Succeed a search even with spelling errors
Hi,
Here, just an implementation of the Levenshtein algorithm.
https://www.php.net/manual/en/function.levenshtein.php

The goal is to find some products with the search functionality. Even if you get some miss spelling in the search input.
That works perfectly with the vanilla Prestashop with default product.

For example, you search  "ummi" instead of "hummi", search controller will be able to find the close word, here "hummingbird", and will display all hummingbird products in the result.

Some precisions about the changes:
-------------------------------------------------
a. In case there are many close words, we take the word with the heaviest word of the list.
b. The algorithm is really fast, only 0.005 - 0.010 second to check 2000 words.
b. The algorithm could find some words, really far for the original.
c. Later in a another commit, we could improvement again the script, for example, by adding the possibility to modify the coefficients of the length of words we want to compare in the BO.

Regards,
Lathanao