Replaced product-flags class' width from 100% to auto
Fixes a regression that made a product miniature non clickable on the right side of a product flag like 'new' or '-20%'. The product-flags class' width was set to 100%, making the link to product non reachable on this div width. It is now set to auto so the div width in not wider thant the product flag.

Resolves #15260