Cart total quantity refresh issue #16665
Modified the object that is sent when the `updateCart` is emitted.
Added an element with key `resp` and value empty object. The empty object is because `resp` is already included in dataset before, so we only need the key to avoid undefined.
Now the situation is the same accross all javascript files , whenever the event `updateCart` is emitted,  the same object keys are sent, whether from `/core/_cart.js` or from `/_dev/cart.js