Avoiding to call several times the actionProductUpdate hook
If we save or update a product with any combination, the for bucle that calls procdessProductAttribute will finally execute the line Hook::exec('actionProductUpdate', array('id_product' => (int) $product->id, 'product' => $product)); one time per each combination. It can be understood as a great performance problem if the user has a very big module that, for example, regenerates an index per each call.

The same Hook is also called in update and uptadeWS Product's functions, I do not know if could be better to just delete the line.