Improve convertPrice function in Tools.php
I've had to resort to this function to be able to format a price and compare it. Unlike other similar functions like displayPrice, here it doesn't  validate if a price is numeric but it comes as a string (something that happens on many occasions). With this little validation we make sure that whenever we enter a valid value, the result will be returned correctly.

On the other hand, I have removed the comment from @deprecated, I understand that until the function is available, this comment should not be displayed because it misleads the developer.

https://github.com/PrestaShop/PrestaShop/issues/14956