Refactor Tax calculation with VAT numbers
Previously, the check for the vatnumbers module was done in several
different places. This change adds a single method on the Tax class that
deteremines the following criteria:
- Is the VAT number specified?
- Is VAT management even enabled?
- Is the VAT from another country than the merchant?

This refactor has two main benefits:
- Code is clearer and more readable, and the concern of the VAT
is in the Tax class, where it belongs;
- Now this is trivially overriddable, for example it allows to perform a
VIES check on the fly if necessary, and not only on data-entry.