Bug searching for alias table into no_quotes array
The no_quotes array contains the execution of the AbstractProcessor::revokeQuotation function from the PHPSQLParser composer package. This function returns an array where it is included all the parts, not a string, this is why it is required to navigate also to the parts array and get the first index.