add check for visitor's country
The visitor's country needs to be checked before the default country is used in the event that Geolocation is on. This then selects the proper carriers for the visitors country and properly displays them in the shopping according to the options set in Shipping->preferences.

iso_code_country is used to create the country_id for the visitor because the visitor's cookie does not contain id_country.

This is related to https://github.com/PrestaShop/PrestaShop/pull/13517.