Force lowercase email addresses to fix missing variables in templates
If the customer gave an email address with uppercase characters (eg `foo@bar.Com` instead of `foo@bar.com`), the `templateVars` are passed to Swift in an array indexed by the given email (eg `foo@bar.Com`) here:

```
$swift->registerPlugin(new \Swift_Plugins_DecoratorPlugin(array($toPlugin => $templateVars)));
```

Problem is: Swift looks for the lowercase key (eg `foo@bar.com`) and doesn't find variables. We need to force lowercase on email addresses to fix this.