Leave the id unchaged if forced_id is set to true
As a developer I would like to be able to extend this class and have the id unchaged if I define forced_id to true.
The same way that the id is not deleted before the insert, it shouldn't be set to 0 once the insert has taken place.

In MySql, when the row id is not generated by the autoincrement mechanism and is set in the insert query the Db::getInstance()->Insert_ID() statement returns 0. That's why I think that it should not be assigned to $this->id when $this_>force_id is set to true.