Use jquery pseudo selector
I'm not sure whether this is a general problem, but in the theme I am using, the country is in a `select`.
This is not picked up by the current selector, meaning that the country is reset each time it is selected. This means clients cannot add addresses from their account page.
This commit fixes the problem.