Disable "Options -MultiViews" by default
Leaving this option active will cause an HTTP error 500 on shared hostings when attempting to upgrade using the "1-Click Upgrade" plugin.
Due to this fact, it is not possible to use the plugin for a seamless update. To perform the update correctly, the update archive must be downloaded and modified and the store updated from it.