Db: add support for SQL transactions
Transactions should be used when executing multiple related queries.
They are required to keep database in consistent state in case of any
errors.

MySQL supports transactions only when using InnoDB however this code is
safe even for MyISAM. PHP "checks for transaction capabilities" for a
given driver only. So this code won't cause exception on MyISAM (it'll
just do nothing).

Ref: https://www.php.net/manual/en/pdo.transactions.php