Mail: add PS_MAIL_SUBJECT_PREFIX to control [Foo] subject prefix
Mail::send() was prefixing every e-mail subject with a shop name in
square brackets. A lot of shop owners look for a way of disabling that
behaviour. Most of them probably find it redundant since there is
already a sender in the From field.

Add PS_MAIL_SUBJECT_PREFIX config option and allow changing it in the
Advanced Parameters → E-mail

Make it enabled by default to avoid changing a long-known behaviour.

Signed-off-by: Rafał Miłecki <rafal@milecki.pl>