Update sources.html.twig
On backend sources, it looks nasty bottom of the page. Because the header takes over half of the page (col-md-6) and the other half is for sources. I think you can use card-header and card-body classes. If its possible also can you add scroll for card-body for sources ( something like: overflow: auto; max-height: 40vh;). If source is long then admin order page is so big.