Update db_structure.sql because it's cleaner
Update db_structure.sql because it's cleaner and clear what what is. Siret = Siret Ape =Ape  vat_number = vat_number  DNI=DNI and IBAN=IBAN and don't sell prestashop short by stuffing it all into DNI. or Siret or even vat into siret??? No let''s keep a clean shop and don't put it into a box in the corner of the shop. Because the are people and corporations also looking and the see it or see the box's in the shop laying around.