FIX on upload attachments file with POST method
I was trying to implement the attachments webservices endpoints and I encountered an error when I do a POST method with a document through  https://<url>/api/attachments/file/ : <![CDATA[[PHP Warning #2] Trying to access array offset on value of type null (/var/www/prestashop/classes/webservices/WebservicesSpecificManagementAttachments.php, line 344)]]>

But the document is upload on the server and is available on files sections on Prestshop BO.

After some analysis and tests I found the origine of the issue, name is nullable. And the attribute name is null when you post your file. With this fix, the upload response contains the good attachment object and I can continue to work with the file if i want to change the name or some attributes.

This modification was introduce with this commit : https://github.com/PrestaShop/PrestaShop/commit/8731bab209a580f9248817a0ebf42028e5ba9774