Follow symlinks for traductions - Deployer best practice
The Finder used to search traduction files does not follow symlinks by default. Which is a major problem when working with modern and robust solutions like Deployer.

On those environments, each GIT release is in a specific folder, and the "living" files (traductions, images...) are in a common shared folder. And symlinks are made to point to the shared folder.

So for now, all the traduction files are ignored by Presta, as they are symlinks. This PR aims to integrate them