Removes outdated config_XX.xml to force module version update
This change forces the deletion of the localized configuration file (config_XX.xml, where XX is the ISO code of the current language, e.g., it for Italian), if it exists, during the module migration.

It was introduced to solve an issue encountered in specific cases:
when the config_it.xml file (or the equivalent for other languages) exists and is more recent than the module's main PHP file ($name.php), the system considers the XML file up to date and does not regenerate it, thus preventing the related database record from being updated.