Update _nav_bar.scss
Bug Fix - Menu on mobile not working clickable
When installing the new version 8.2.0.
After you open the administration on mobile and click on the burger menu. The menu will slide out, but it is completely covered by an overlay, so it is not clickable.
Thanks to this change, the menu is working again.